requirements.txt

oauth2client==4.1.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
google-api-python-client==2.108.0